task-updating

See merge request big-d/monorepo!29

+++ apps/goal-service/src/modules/tasks/application/use-cases/complete-delete-task/complete-delete-task.use-case.ts
@@ -1,10 +1,13 @@
-import { ExceptionTaskNotFound } from '@/modules/tasks/application/exceptions';
+import {
+  ExceptionTaskNotFound,
+  ExceptionTaskUnprocessable,
+} from '@/modules/tasks/application/exceptions';
 import { TaskDatabase, TasksWriteRepository } from '@/modules/tasks/application/ports';
 import { TaskFactory } from '@/modules/tasks/domain';
 import { TasksToken } from '@/modules/tasks/tokens';
 import { databaseToken } from '@big-d/database';
 import { Inject, Injectable } from '@nestjs/common';
-import { TaskCheckerService } from '../../services';
+import { TaskCheckerService, TaskTypeService } from '../../services';
 import { CompleteDeleteTaskCommand } from './complete-delete-task.command';
 
 @Injectable()
@@ -12,27 +15,34 @@ class CompleteDeleteTaskUseCase {
   constructor(
     @Inject(TasksToken.WRITE_REPOSITORY) private readonly tasksWriteRepo: TasksWriteRepository,
     @Inject(databaseToken.CONNECTION) private readonly db: TaskDatabase,
+    private readonly taskTypeService: TaskTypeService,
 
     private readonly taskCheckerService: TaskCheckerService,
   ) {}
 
   async execute({ input }: CompleteDeleteTaskCommand): Promise<{ id: number }> {
     return this.db.runTransaction(async (trx) => {
-      const task = await this.taskCheckerService.ensureTaskExists(input, { trx });
-      TaskFactory.deleteComplete(task);
+      const { taskId, userId } = input;
+      const { isOrigin, data } = this.taskTypeService.getType({ taskId });
 
-      const isDeleted = await this.tasksWriteRepo.deleteTask(
-        { taskId: input.taskId, userId: input.userId },
-        trx,
-      );
+      if (isOrigin) {
+        const task = await this.taskCheckerService.ensureTaskExists(
+          { taskId: data.id, userId },
+          { trx },
+        );
+        TaskFactory.deleteComplete(task);
 
-      if (!isDeleted) {
-        throw new ExceptionTaskNotFound({ taskId: input.taskId });
+        const isDeleted = await this.tasksWriteRepo.deleteTask({ taskId: data.id, userId }, trx);
+        if (!isDeleted) {
+          throw new ExceptionTaskNotFound({ taskId: data.id });
+        }
+
+        return {
+          id: task.id,
+        };
       }
 
-      return {
-        id: input.taskId,
-      };
+      throw new ExceptionTaskUnprocessable({ taskId, message: 'Не валидный id' });
     });
   }
 }
